[FEAT] API 경로 수정
