+++ src/main/java/com/cone/cone/domain/user/controller/MenteeApi.java
@@ -1,5 +1,6 @@
 package com.cone.cone.domain.user.controller;
 
+import com.cone.cone.domain.mentorings.dto.response.*;
 import com.cone.cone.domain.room.entity.Room;
 import com.cone.cone.domain.user.dto.response.*;
 import com.cone.cone.global.response.ResponseTemplate;
@@ -11,4 +12,6 @@ public interface MenteeApi {
     ResponseEntity<ResponseTemplate<List<Room>>> getRoomsById(Long id);
 
     ResponseEntity<ResponseTemplate<MenteeProfileResponse>> getMenteeProfile(Long menteeId);
+
+    ResponseEntity<ResponseTemplate<List<MenteeMentoringResponse>>> getMentoringsForMentee(Long menteeId);
 }
